Convert int to long to prevent overflow

+++ tlatools/src/tlc2/tool/fp/DiskFPSet.java
@@ -301,7 +301,7 @@ public class DiskFPSet extends FPSet implements FPSetStatistic {
 			for (int i = 0; i < this.tbl.length; i++) {
 				if (this.tbl[i] != null) {
 					// 16 bytes overhead for each row in tbl!
-					size += 16 + (this.tbl[i].length * LongSize);
+					size += 16 + (this.tbl[i].length * (long) LongSize);
 				}
 			}
 			// size of index array if non-null
